refactor: task list project module refactor
