+++ src/task-list-project/task-list-project.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { TaskListProjectController } from './controller/task-list-project/task-list-project.controller';
-import { TaskListProjectService } from './service/task-list-project/task-list-project.service';
+import { TaskListProjectController } from './controller/task-list-project.controller';
+import { TaskListProjectService } from './service/task-list-project.service';
 import { SharedServicesModule } from 'src/shared/shared-services.module';
 
 @Module({
